Traffic around structures https://jira.ccpgames.com/browse/GFX-596
Addendum to CL 967484

Binaries submitted in 967516
Build time : 9 minutes 16.779000 seconds
Branch : //eve/branches/development/MAINLINE/

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/Trinity/": change = 967507]

+++ Tr2RuntimeInstanceData.cpp
@@ -43,7 +43,7 @@ void Tr2RuntimeInstanceData::SaveToGranny( const char* resPath ) const
 	std::wstring filename = BePaths->ResolvePathForWritingW( resPathW );
 
 	granny_data_type_definition* definition = CCP_NEW( "Tr2RuntimeInstanceData::SaveToGranny" ) granny_data_type_definition[m_layout.m_items.size() + 1];
-	if( !ConvertVertexDeclToGranny( m_layout, definition, m_layout.m_items.size() ) )
+	if( !ConvertVertexDeclToGranny( m_layout, definition, (unsigned int)( m_layout.m_items.size() ) ) )
 	{
 		return;
 	}
